ev3/snippets/bluetooth_pc: run MicroPython Unix
This is a copy of the EV3 server example with minimal changes
to make it run on a PC, using the unmodified MicroPython Unix port.

This is done by skipping the name resolver. The downside is that you
must specify the EV3 server address instead of its name.  However,
since in this example the PC will be the server, this is not really a
problem. The EV3 clients will just connect to the PC by its name.