ev3/snippets/bluetooth_pc: PC is client
Technically it does not matter much who is the server or the client.

However, if Windows PCs are involved, we found it to be easier if
Windows is the client and EV3 is the server.
That way, it is the Windows PC initiating the connection. It doesn't
seem to like it when other devices take the initiative.

So, we update the example accordingly to make it
accessible for most users without loss of generality.

You can still make the PC the server by just running the
server example on your PC.