official_models/pup/trains: passenger and cargo
These two are likely the first in a range of generic LEGO trains with a
City Hub and a DC Train Motor. Instead of adding examples to each,
we add generic train examples in a shared folder and link to them.

We reserve these folders specifically for these sets and their B models.